фикс

+++ internal/domain/models/accrual.go
@@ -3,5 +3,5 @@ package models
 type Accrual struct {
 	OrderNum string      `json:"order"`
 	Status   OrderStatus `json:"status"`
-	Accrual  int64       `json:"accrual"`
+	Accrual  float64     `json:"accrual"`
 }
